+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/core/data/local/FavoriteCameraEntity.kt
@@ -1,0 +1,6 @@
+package com.sanathcoding.sglivetrafficimage.core.data.local
+
+data class FavoriteCameraEntity(
+    val cameraId: String,
+    val userName: String
+)
