+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/core/data/local/UserCredentialEntity.kt
@@ -1,0 +1,11 @@
+package com.sanathcoding.sglivetrafficimage.core.data.local
+
+import androidx.room.Entity
+import androidx.room.PrimaryKey
+
+@Entity
+data class UserCredentialEntity(
+    val userName: String,
+    val password: String,
+    @PrimaryKey val id:Int? = null
+)
