+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/core/navigation/Screen.kt
@@ -1,0 +1,9 @@
+package com.sanathcoding.sglivetrafficimage.core.navigation
+
+sealed class Screen(val route: String) {
+    object Login: Screen("login")
+    object Home: Screen("home")
+    object CameraDetail: Screen("camera_detail/{cameraId}") {
+        fun createRoute(cameraId: String) = "camera_detail/$cameraId"
+    }
+}
