+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/core/navigation/SetUpNavGraph.kt
@@ -1,0 +1,33 @@
+package com.sanathcoding.sglivetrafficimage.core.navigation
+
+import androidx.compose.runtime.Composable
+import androidx.navigation.NavHostController
+import androidx.navigation.compose.NavHost
+import androidx.navigation.compose.composable
+import com.sanathcoding.sglivetrafficimage.presentation.camera_details_screen.CameraDetailScreen
+import com.sanathcoding.sglivetrafficimage.presentation.home_screen.NewHome
+import com.sanathcoding.sglivetrafficimage.feature_login.presentation.loginScreen.LoginScreen
+
+const val TEST_ROUTE = "test_route"
+
+@Composable
+fun SetupNavGraph(navController: NavHostController) {
+    NavHost(
+        navController = navController,
+        route = TEST_ROUTE,
+        startDestination = Screen.Login.route
+    ) {
+        composable(Screen.Login.route) {
+            LoginScreen(navController = navController)
+        }
+        composable(Screen.Home.route) {
+            NewHome(navController)
+        }
+        composable(Screen.CameraDetail.route) { naveBackStackEntry ->
+            val cameraId = naveBackStackEntry.arguments?.getString("cameraId")
+            cameraId?.let {
+                CameraDetailScreen(it)
+            }
+        }
+    }
+}
