--- app/src/main/java/com/sanathcoding/sglivetrafficimage/data/local/FavoriteCameraEntity.kt
@@ -1,6 +1,0 @@
-package com.sanathcoding.sglivetrafficimage.data.local
-
-data class FavoriteCameraEntity(
-    val cameraId: String,
-    val userName: String
-)
