--- app/src/main/java/com/sanathcoding/sglivetrafficimage/data/local/UserCredentialDataBase.kt
@@ -1,13 +1,0 @@
-package com.sanathcoding.sglivetrafficimage.data.local
-
-import androidx.room.Database
-import androidx.room.RoomDatabase
-
-@Database(
-    entities = [UserCredentialEntity::class],
-    version = 1,
-    exportSchema = false
-)
-abstract class UserCredentialDataBase: RoomDatabase() {
-    abstract val dao: UserCredentialDao
-}
