--- app/src/main/java/com/sanathcoding/sglivetrafficimage/data/local/UserCredentialEntity.kt
@@ -1,11 +1,0 @@
-package com.sanathcoding.sglivetrafficimage.data.local
-
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-
-@Entity
-data class UserCredentialEntity(
-    val userName: String,
-    val password: String,
-    @PrimaryKey val id:Int? = null
-)
