--- app/src/main/java/com/sanathcoding/sglivetrafficimage/domain/model/UserCredential.kt
@@ -1,7 +1,0 @@
-package com.sanathcoding.sglivetrafficimage.domain.model
-
-data class UserCredential(
-    val userName: String,
-    val password: String,
-    val id: Int? = null
-)
