+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/feature_login/domain/model/UserCredential.kt
@@ -1,0 +1,7 @@
+package com.sanathcoding.sglivetrafficimage.feature_login.domain.model
+
+data class UserCredential(
+    val userName: String,
+    val password: String,
+    val id: Int? = null
+)
