+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/feature_login/domain/repository/UserCredentialRepository.kt
@@ -1,0 +1,28 @@
+package com.sanathcoding.sglivetrafficimage.feature_login.domain.repository
+
+import com.sanathcoding.sglivetrafficimage.core.data.local.UserCredentialEntity
+import com.sanathcoding.sglivetrafficimage.feature_login.domain.model.UserCredential
+import kotlinx.coroutines.flow.Flow
+
+interface UserCredentialRepository {
+
+    suspend fun insertUser(user: UserCredential)
+
+//    suspend fun getIsUserNameAvailable(name: String): String
+
+//    fun getUserByUserName(name: String): UserCredentialEntity
+
+    fun doseUserExist(name: String): Int
+
+
+    suspend fun getPasswordByUserName(name: String): UserCredentialEntity
+
+//    suspend fun getPasswordByUserName(name: String): String
+
+    fun getUser(): Flow<List<UserCredential>>
+
+    suspend fun putString(key: String, value: String)
+
+    suspend fun getString(key: String): String?
+
+}
