+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/feature_login/domain/util/validationResult.kt
@@ -1,0 +1,8 @@
+package com.sanathcoding.sglivetrafficimage.feature_login.domain.util
+
+import com.sanathcoding.sglivetrafficimage.core.util.UiText
+
+data class ValidationResult(
+    val isSuccessful: Boolean,
+    val errorMsg: UiText? = null
+)
