+++ app/src/main/java/com/sanathcoding/sglivetrafficimage/presentation/home_screen/screen/CameraListScreen.kt
@@ -9,7 +9,7 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.navigation.NavHostController
-import com.sanathcoding.sglivetrafficimage.navigation.Screen
+import com.sanathcoding.sglivetrafficimage.core.navigation.Screen
 
 @Composable
 fun CameraListScreen(
