--- app/src/main/java/com/sanathcoding/sglivetrafficimage/presentation/loginScreen/ValidateEvent.kt
@@ -1,7 +1,0 @@
-package com.sanathcoding.sglivetrafficimage.presentation.loginScreen
-
-sealed interface ValidateEvent {
-    object Success: ValidateEvent
-    object DataAdded: ValidateEvent
-    object PasswordNotMatch: ValidateEvent
-}
